requirements.txt
customtkinter==5.2.2
Pillow==11.1.0
python-dotenv==1.0.1
Requests==2.32.3
